src/models/requirements-serve.txt
Pillow
numpy
flask
pandas
opencv-python